requirements.txt
dj-database-url==2.1.0
Django==3.2.7
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9


django-redis==5.4.0
redis==5.0.1
gunicorn==21.2.0
whitenoise==6.6.0